requirements.txt
transformers
langchain
chroma
ragas
torch
scikit-learn
numpy
pandas
spacy
fastapi
uvicorn
elasticsearch
faiss-cpu
pymongo
requests
Jinja2
python-dotenv
cohere
deeplake
tiktoken
langchain_cohere
docx2txt
pypdf
openai